# API_Doc.yaml
openapi: 3.0.2
info:
  title: Fast Report API
  description: Хранилище данных
  version: 0.1.0
paths:
  /fif/create_folder:
    post:
      tags:
        - Work with files and folders
      summary: Create Folder
      operationId: create_folder_fif_create_folder_post
      parameters:
        - required: false
          schema:
            allOf:
              - $ref: '#/components/schemas/SystemItemTag'
            default: Other
          name: tag
          in: query
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FolderItemRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemData'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /fif/create_file:
    post:
      tags:
        - Work with files and folders
      summary: Create File
      operationId: create_file_fif_create_file_post
      parameters:
        - required: false
          schema:
            title: Parent Id
            type: integer
          name: parent_id
          in: query
        - required: false
          schema:
            allOf:
              - $ref: '#/components/schemas/SystemItemTag'
            default: Other
          name: tag
          in: query
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_create_file_fif_create_file_post'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemData'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /fif/create_file_by_path:
    post:
      tags:
        - Work with files and folders
      summary: Create File By Path
      operationId: create_file_by_path_fif_create_file_by_path_post
      parameters:
        - required: false
          schema:
            allOf:
              - $ref: '#/components/schemas/SystemItemTag'
            default: Other
          name: tag
          in: query
        - required: false
          schema:
            title: File Url
            type: string
            default: data/files/
          name: file_url
          in: query
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_create_file_by_path_fif_create_file_by_path_post'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemData'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /fif/delete/{id}:
    delete:
      tags:
        - Work with files and folders
      summary: Delete
      operationId: delete_fif_delete__id__delete
      parameters:
        - required: true
          schema:
            title: Item Id
            type: integer
          name: item_id
          in: query
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /fif/delete_by_url:
    delete:
      tags:
        - Work with files and folders
      summary: Delete By Url
      operationId: delete_by_url_fif_delete_by_url_delete
      parameters:
        - required: true
          schema:
            title: Url
            type: string
          name: url
          in: query
        - required: true
          schema:
            title: Date
            type: string
            format: date-time
          name: date
          in: query
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /fif/get_by_tag:
    get:
      tags:
        - Work with files and folders
      summary: Get Files By Tag
      operationId: get_files_by_tag_fif_get_by_tag_get
      parameters:
        - required: false
          schema:
            allOf:
              - $ref: '#/components/schemas/SystemItemTag'
            default: Other
          name: tag
          in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemTagResponseData'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /fif/get_by_tag_and_path:
    get:
      tags:
        - Work with files and folders
      summary: Get Files By Tag
      operationId: get_files_by_tag_fif_get_by_tag_and_path_get
      parameters:
        - required: false
          schema:
            allOf:
              - $ref: '#/components/schemas/SystemItemTag'
            default: Other
          name: tag
          in: query
        - required: false
          schema:
            title: File Url
            type: string
            default: data/files/
          name: file_url
          in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemTagResponseData'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /fif/nodes/{id}:
    get:
      tags:
        - Work with files and folders
      summary: Get Files
      operationId: get_files_fif_nodes__id__get
      parameters:
        - required: true
          schema:
            title: Id
            type: integer
          name: id
          in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResponseData'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_create_file_by_path_fif_create_file_by_path_post:
      title: Body_create_file_by_path_fif_create_file_by_path_post
      required:
        - file
      type: object
      properties:
        file:
          title: File
          type: string
          format: binary
    Body_create_file_fif_create_file_post:
      title: Body_create_file_fif_create_file_post
      required:
        - file
      type: object
      properties:
        file:
          title: File
          type: string
          format: binary
...
    FolderItemRequest:
      title: FolderItemRequest
      required:
        - name
      type: object
      properties:
        parent_id:
          title: Parent Id
          type: integer
        name:
          title: Name
          type: string
...
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        detail:
          title: Detail
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
    ItemData:
      title: ItemData
      required:
        - id
        - date
        - size
        - url
        - name
        - tag
        - type
      type: object
      properties:
        id:
          title: Id
          type: integer
        date:
          title: Date
          type: string
          format: date-time
        parent_id:
          title: Parent Id
          type: integer
        size:
          title: Size
          type: integer
        url:
          title: Url
          type: string
        name:
          title: Name
          type: string
        tag:
          $ref: '#/components/schemas/SystemItemTag'
        type:
          $ref: '#/components/schemas/SystemItemType'
    ItemResponseData:
      title: ItemResponseData
      required:
        - id
        - date
        - size
        - url
        - name
        - tag
        - type
      type: object
      properties:
        id:
          title: Id
          type: integer
        date:
          title: Date
          type: string
          format: date-time
        parent_id:
          title: Parent Id
          type: integer
        size:
          title: Size
          type: integer
        url:
          title: Url
          type: string
        name:
          title: Name
          type: string
        tag:
          $ref: '#/components/schemas/SystemItemTag'
        type:
          $ref: '#/components/schemas/SystemItemType'
        children:
          title: Children
          type: array
          items: {}
          default: []
    ItemTagResponseData:
      title: ItemTagResponseData
      required:
        - items
      type: object
      properties:
        items:
          title: Items
          type: array
          items:
            $ref: '#/components/schemas/ItemData'
    SystemItemTag:
      title: SystemItemTag
      enum:
        - Document
        - Template
        - Report
        - Export
        - Other
      type: string
      description: An enumeration.
    SystemItemType:
      title: SystemItemType
      enum:
        - FILE
        - FOLDER
      type: string
      description: An enumeration.
...
    ValidationError:
      title: ValidationError
      required:
        - loc
        - msg
        - type
      type: object
      properties:
        loc:
          title: Location
          type: array
          items:
            anyOf:
              - type: string
              - type: integer
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string